12_REST_API_DJANGO/requirements.txt
Django==4.0.6
djangorestframework==3.13.1
